requirements.txt
google-genai
aiokafka
kafka-python-ng
redis[hiredis]
ddgs
crawl4ai
python-dotenv
streamlit
